+++ providers/dns/joker/joker.go
@@ -1,17 +1,14 @@
-// Package joker implements a DNS provider for solving the DNS-01 challenge using joker.com DMAPI.
+// Package joker implements a DNS provider for solving the DNS-01 challenge using joker.com.
 package joker
 
 import (
-	"errors"
-	"fmt"
 	"net/http"
-	"strings"
+	"os"
 	"time"
 
+	"github.com/go-acme/lego/v4/challenge"
 	"github.com/go-acme/lego/v4/challenge/dns01"
-	"github.com/go-acme/lego/v4/log"
 	"github.com/go-acme/lego/v4/platform/config/env"
-	"github.com/go-acme/lego/v4/providers/dns/joker/internal/dmapi"
 )
 
 // Environment variables names.
@@ -22,11 +19,17 @@ const (
 	EnvUsername = envNamespace + "USERNAME"
 	EnvPassword = envNamespace + "PASSWORD"
 	EnvDebug    = envNamespace + "DEBUG"
+	EnvMode     = envNamespace + "API_MODE"
 
 	EnvTTL                = envNamespace + "TTL"
 	EnvPropagationTimeout = envNamespace + "PROPAGATION_TIMEOUT"
 	EnvPollingInterval    = envNamespace + "POLLING_INTERVAL"
 	EnvHTTPTimeout        = envNamespace + "HTTP_TIMEOUT"
+)
+
+const (
+	modeDMAPI = "DMAPI"
+	modeSVC   = "SVC"
 )
 
 // Config is used to configure the creation of the DNSProvider.
@@ -35,6 +38,7 @@ type Config struct {
 	APIKey             string
 	Username           string
 	Password           string
+	APIMode            string
 	PropagationTimeout time.Duration
 	PollingInterval    time.Duration
 	TTL                int
@@ -44,6 +48,7 @@ type Config struct {
 // NewDefaultConfig returns a default configuration for the DNSProvider.
 func NewDefaultConfig() *Config {
 	return &Config{
+		APIMode:            env.GetOrDefaultString(EnvMode, modeDMAPI),
 		Debug:              env.GetOrDefaultBool(EnvDebug, false),
 		TTL:                env.GetOrDefaultInt(EnvTTL, dns01.DefaultTTL),
 		PropagationTimeout: env.GetOrDefaultSecond(EnvPropagationTimeout, dns01.DefaultPropagationTimeout),
@@ -54,152 +59,21 @@ func NewDefaultConfig() *Config {
 	}
 }
 
-// DNSProvider implements the challenge.Provider interface.
-type DNSProvider struct {
-	config *Config
-	client *dmapi.Client
-}
-
-// NewDNSProvider returns a DNSProvider instance configured for Joker DMAPI.
+// NewDNSProvider returns a DNSProvider instance configured for Joker.
 // Credentials must be passed in the environment variable JOKER_API_KEY.
-func NewDNSProvider() (*DNSProvider, error) {
-	values, err := env.Get(EnvAPIKey)
-	if err != nil {
-		var errU error
-		values, errU = env.Get(EnvUsername, EnvPassword)
-		if errU != nil {
-			return nil, fmt.Errorf("joker: %v or %v", errU, err)
-		}
+func NewDNSProvider() (challenge.ProviderTimeout, error) {
+	if os.Getenv(EnvMode) == modeSVC {
+		return newSvcProvider()
 	}
 
-	config := NewDefaultConfig()
-	config.APIKey = values[EnvAPIKey]
-	config.Username = values[EnvUsername]
-	config.Password = values[EnvPassword]
-
-	return NewDNSProviderConfig(config)
+	return newDmapiProvider()
 }
 
-// NewDNSProviderConfig return a DNSProvider instance configured for Joker DMAPI.
-func NewDNSProviderConfig(config *Config) (*DNSProvider, error) {
-	if config == nil {
-		return nil, errors.New("joker: the configuration of the DNS provider is nil")
+// NewDNSProviderConfig return a DNSProvider instance configured for Joker.
+func NewDNSProviderConfig(config *Config) (challenge.ProviderTimeout, error) {
+	if config.APIMode == modeSVC {
+		return newSvcProviderConfig(config)
 	}
 
-	if config.APIKey == "" {
-		if config.Username == "" || config.Password == "" {
-			return nil, errors.New("joker: credentials missing")
-		}
-	}
-
-	client := dmapi.NewClient(dmapi.AuthInfo{
-		APIKey:   config.APIKey,
-		Username: config.Username,
-		Password: config.Password,
-	})
-
-	client.Debug = config.Debug
-
-	if config.HTTPClient != nil {
-		client.HTTPClient = config.HTTPClient
-	}
-
-	return &DNSProvider{config: config, client: client}, nil
-}
-
-// Timeout returns the timeout and interval to use when checking for DNS propagation.
-func (d *DNSProvider) Timeout() (timeout, interval time.Duration) {
-	return d.config.PropagationTimeout, d.config.PollingInterval
-}
-
-// Present installs a TXT record for the DNS challenge.
-func (d *DNSProvider) Present(domain, token, keyAuth string) error {
-	fqdn, value := dns01.GetRecord(domain, keyAuth)
-
-	zone, err := dns01.FindZoneByFqdn(fqdn)
-	if err != nil {
-		return fmt.Errorf("joker: %w", err)
-	}
-
-	relative := getRelative(fqdn, zone)
-
-	if d.config.Debug {
-		log.Infof("[%s] joker: adding TXT record %q to zone %q with value %q", domain, relative, zone, value)
-	}
-
-	response, err := d.client.Login()
-	if err != nil {
-		return formatResponseError(response, err)
-	}
-
-	response, err = d.client.GetZone(zone)
-	if err != nil || response.StatusCode != 0 {
-		return formatResponseError(response, err)
-	}
-
-	dnsZone := dmapi.AddTxtEntryToZone(response.Body, relative, value, d.config.TTL)
-
-	response, err = d.client.PutZone(zone, dnsZone)
-	if err != nil || response.StatusCode != 0 {
-		return formatResponseError(response, err)
-	}
-
-	return nil
-}
-
-// CleanUp removes a TXT record used for a previous DNS challenge.
-func (d *DNSProvider) CleanUp(domain, token, keyAuth string) error {
-	fqdn, _ := dns01.GetRecord(domain, keyAuth)
-
-	zone, err := dns01.FindZoneByFqdn(fqdn)
-	if err != nil {
-		return fmt.Errorf("joker: %w", err)
-	}
-
-	relative := getRelative(fqdn, zone)
-
-	if d.config.Debug {
-		log.Infof("[%s] joker: removing entry %q from zone %q", domain, relative, zone)
-	}
-
-	response, err := d.client.Login()
-	if err != nil {
-		return formatResponseError(response, err)
-	}
-
-	defer func() {
-		// Try to logout in case of errors
-		_, _ = d.client.Logout()
-	}()
-
-	response, err = d.client.GetZone(zone)
-	if err != nil || response.StatusCode != 0 {
-		return formatResponseError(response, err)
-	}
-
-	dnsZone, modified := dmapi.RemoveTxtEntryFromZone(response.Body, relative)
-	if modified {
-		response, err = d.client.PutZone(zone, dnsZone)
-		if err != nil || response.StatusCode != 0 {
-			return formatResponseError(response, err)
-		}
-	}
-
-	response, err = d.client.Logout()
-	if err != nil {
-		return formatResponseError(response, err)
-	}
-	return nil
-}
-
-func getRelative(fqdn, zone string) string {
-	return dns01.UnFqdn(strings.TrimSuffix(fqdn, dns01.ToFqdn(zone)))
-}
-
-// formatResponseError formats error with optional details from DMAPI response.
-func formatResponseError(response *dmapi.Response, err error) error {
-	if response != nil {
-		return fmt.Errorf("joker: DMAPI error: %w Response: %v", err, response.Headers)
-	}
-	return fmt.Errorf("joker: DMAPI error: %w", err)
+	return newDmapiProviderConfig(config)
 }
